feat(openai): Add Driver class to endpoint module
This commit adds the `Driver` class to the `endpoint` module in the `sdk` package. The `Driver` class represents the configuration for the OpenAI endpoint. It includes properties such as the endpoint URL, API key, organization ID, and model. The `Driver` class also provides methods for checking the availability of the configuration and generating a unique identifier.

The `Driver` class is used in various parts of the codebase, such as in the `llm_provider` and `service_provider` modules, to request and authenticate the OpenAI driver.

Note: This commit only introduces the class and does not make any changes to its implementation.
