+++ llmkira/sdk/endpoint/openai/__init__.py
@@ -5,16 +5,15 @@
 # @Software: PyCharm
 __version__ = "0.0.1"
 
-import hashlib
-import os
 from typing import Union, List, Optional, Literal
 
 import httpx
 from dotenv import load_dotenv
 from loguru import logger
-from pydantic import BaseModel, root_validator, validator, Field, HttpUrl, BaseSettings
+from pydantic import BaseModel, root_validator, validator, Field, BaseSettings
 
 from .action import Tokenizer, TokenizerObj
+from ..tee import Driver
 from ...error import ValidationError
 from ...network import request
 from ...schema import Message, Function
@@ -36,16 +35,6 @@ MODEL = Literal[
     # "gpt-4-32k-0314"
     # do not use 0314. check: https://platform.openai.com/docs/guides/gpt/function-calling
 ]
-
-
-def sha1_encrypt(string):
-    """
-    sha1加密算法
-    """
-
-    sha = hashlib.sha1(string.encode('utf-8'))
-    encrypts = sha.hexdigest()
-    return encrypts[:8]
 
 
 class OpenaiResult(BaseModel):
@@ -90,95 +79,6 @@ class Openai(BaseModel):
             env_file_encoding = 'utf-8'
             case_sensitive = True
             arbitrary_types_allowed = True
-
-    class Driver(BaseSettings):
-        """
-        不允许部分更新
-        """
-        endpoint: HttpUrl = Field(default="https://api.openai.com/v1/chat/completions")
-        api_key: str = Field(default=None)
-        org_id: Optional[str] = Field(None)
-        model: MODEL = Field(default="gpt-3.5-turbo-0613")
-        model_retrieve: MODEL = Field(default="gpt-3.5-turbo-16k")
-
-        # TODO:AZURE API VERSION
-        @property
-        def detail(self):
-            """
-            脱敏
-            """
-            api_key = "****" + str(self.api_key)[-4:]
-            return (
-                f"Endpoint: {self.endpoint}\nApi_key: {api_key}\n"
-                f"Org_id: {self.org_id}\nModel: {self.model}\nRetrieve_model: {self.model_retrieve}"
-            )
-
-        @property
-        def available(self):
-            """
-            检查可用性
-            :return:
-            """
-            return all([self.endpoint, self.api_key, self.model])
-
-        @classmethod
-        def from_public_env(cls):
-            openai_api_key = os.getenv("OPENAI_API_KEY", None)
-            openai_endpoint = os.getenv("OPENAI_API_ENDPOINT", "https://api.openai.com/v1/chat/completions")
-            openai_org_id = os.getenv("OPENAI_API_ORG_ID", None)
-            openai_model = os.getenv("OPENAI_API_MODEL", "gpt-3.5-turbo-0613")
-            openai_model_retrieve = os.getenv("OPENAI_API_RETRIEVE_MODEL", "gpt-3.5-turbo-16k")
-            return cls(
-                endpoint=openai_endpoint,
-                api_key=openai_api_key,
-                org_id=openai_org_id,
-                model=openai_model,
-                model_retrieve=openai_model_retrieve
-            )
-
-        @validator("model")
-        def check_model(cls, v):
-            if v not in MODEL.__args__:
-                logger.warning("Validator:model only support {}".format(MODEL.__args__))
-                # raise ValidationError("model only support {}".format(MODEL.__args__))
-            return v
-
-        @validator("model_retrieve")
-        def check_retrieve_model(cls, v):
-            if v not in MODEL.__args__:
-                logger.warning("Validator:model_retrieve only support {}".format(MODEL.__args__))
-                # raise ValidationError("model_retrieve only support {}".format(MODEL.__args__))
-            return v
-
-        @validator("api_key")
-        def check_key(cls, v):
-            if v:
-                if not str(v).startswith("sk-"):
-                    logger.warning("Validator:api_key should start with `sk-`")
-                if len(str(v)) < 4:
-                    raise ValidationError("api_key is too short")
-                # 严格检查
-                # if not len(str(v)) == 51:
-                #    raise ValidationError("api_key must be 51 characters long")
-            else:
-                raise ValidationError("api_key is required,pls set OPENAI_API_KEY in .env")
-            return v
-
-        @property
-        def uuid(self):
-            """
-            取 api key 最后 3 位 和 sha1 加密后的前 8 位
-            :return: uuid for token
-            """
-            _flag = self.api_key[-3:]
-            return f"{_flag}:{sha1_encrypt(self.api_key)}"
-
-        class Config:
-            env_file = '.env'
-            env_file_encoding = 'utf-8'
-            case_sensitive = True
-            arbitrary_types_allowed = True
-            extra = "allow"
 
     messages: List[Message]
     config: Driver
